Move impl_listting_data_connector macro
